dsaf

+++ quanatee/contrib/pentaggregation/aggtrigger/functions/avgl.go
@@ -6,7 +6,7 @@ func AvgLeftFloat32(values []float32) float32 {
 	avg := float32(0)
 	length := len(values)
 	for idx, val := range values {
-		avg += (val/(idx+1))
+		avg += (val/float32(idx+1))
 	}
     
     avg = avg/float32(length)
@@ -19,7 +19,7 @@ func AvgLeftFloat64(values []float64) float64 {
 	length := len(values)
 	for idx, val := range values {
         avg = avg.Add(avg, big.NewFloat(float64(val)))
-        avg = avg.Add(avg, big.NewFloat(float64(val/(idx+1))))
+        avg = avg.Add(avg, big.NewFloat(float64(val/float64(idx+1))))
 	}
 
     avg = new(big.Float).Quo(avg, big.NewFloat(float64(length)))
